Remove MicroPlot.tsx
