+++ packages/grafana-ui/src/components/index.ts
@@ -69,7 +69,6 @@ export {
   BigValueTextMode,
 } from './BigValue/BigValue';
 
-export { MicroPlot } from './uPlot/MicroPlot';
 export { UPlotChart } from './uPlot/Plot';
 export * from './uPlot/plugins';
 export { Gauge } from './Gauge/Gauge';
